Proxy-rewrite: - proxy_url() and proxy_parse_html() are both non-module functions (now   methods) - so they must be splitted into a seperate class - also the SIZE_* and DEFAULT_TIME constants are both not relevant to module
