+++ mod/match.php
@@ -10,7 +10,6 @@ use Friendica\Core\L10n;
 use Friendica\Core\System;
 use Friendica\Database\DBA;
 use Friendica\Model\Contact;
-use Friendica\Module\Proxy;
 use Friendica\Util\Network;
 use Friendica\Util\Proxy as ProxyUtils;
