+++ src/Worker/CronJobs.php
@@ -15,7 +15,6 @@ use Friendica\Model\Contact;
 use Friendica\Model\GContact;
 use Friendica\Model\Photo;
 use Friendica\Model\User;
-use Friendica\Module\Proxy;
 use Friendica\Network\Probe;
 use Friendica\Protocol\PortableContact;
 use Friendica\Util\Proxy as ProxyUtils;
